refactor: move code from file to file
