--- src/modules/models/user.model.ts
@@ -1,6 +1,0 @@
-export interface UserRecord {
-  id?: string;
-  username: string;
-  age: number;
-  hobbies: string[];
-}
